+++ src/main/java/lotto/domain/WinningLottoWithBonus.java
@@ -1,5 +1,7 @@
 package lotto.domain;
 
+import lotto.domain.lotto.Lotto;
+import lotto.domain.lotto.LottoNumber;
 import lotto.dto.PrizeResponse;
 
 import java.util.Arrays;
